Doc: Updated the qdocconf to include .java files
QDoc ignores the .java files without the following entry in the
qdocconf:
examples.fileextensions += *.java

Java files are not part of the default set of file types included
by Qdoc. So there must be an entry in the qdocconf to force Qdoc
to include additional file types such as the .java files.

Task-number: QTBUG-35568
Change-Id: Ic8b805e0dec937fb0e8b19f0b94883ced5eb6c04
Reviewed-by: Christian Stromme <christian.stromme@digia.com>
Reviewed-by: Nico Vertriest <nico.vertriest@digia.com>
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>