Add API for starting intents
We need an API for this, because otherwise it would be
impossible to start intents and get results from them
without having access to the application's activities.
For third-party libraries etc., this is required.

When we define a public Java API, we will implement a similar
API there, so that applications can also start intents from
Java code without worrying about collisions with third-party
add-ons.

[ChangeLog][Intents] Introduced API to launch intents.

Change-Id: Ic3bbfbbaced3278c2ee970e74cba2997d5d867c4
Reviewed-by: Christian Stromme <christian.stromme@digia.com>