Doc: Fix documentation issues when using Clang-enabled QDoc
QDoc in Qt 5.11 uses Clang to parse the C++ source for documentation
comments. Unlike the old parser, Clang requires a module header to
be present as well as correct include paths to generate the docs
correctly. However, these are available from the build system only
when building for Android, not when generating the documentation
on a system with no Android-related assets.

To fix this, create a 'dummy' module header specific to doc builds,
add required relative include paths in .qdocconf, and fix the
documentation \fn commands to work with Clang.

Change-Id: Ia3d4e50ff8756cd8c8aae9cd3c3d1492eb7db328
Reviewed-by: Martin Smith <martin.smith@qt.io>